chore: shrink API surface; merge tach modules; finish Project/Task OO
W4.C completes the API surface and OO cleanup for terok itself, pinning
to the in-flight sandbox + executor PR branches.

Integration adapters reconciled to the upstream W4.A/W4.B trims:
- terok-sandbox top-level shed 11 entries (FakeKrunTransport, KrunTransport,
  TcpEndpoint, DEFAULT_SSH_USER, shield_interactive_session,
  shield_watch_session, is_systemd_creds_available, RawHooksSection,
  GATE_COMMANDS, SSH_COMMANDS, EXIT_MANUAL_STEP_NEEDED) and CommandDef/
  CommandTree now flow through terok_util (their canonical home) rather
  than the sandbox adapter — consistent with the foundation-library rule.
- terok-executor shed apply_provider_config and build_headless_command,
  which are now AgentProvider methods; the headless task runner calls
  them on the resolved provider.

Three tach modules collapsed (~120 LoC net deletion):
- util/text_wrap → ui_utils/terminal (display utilities together)
- orchestration/autopilot → orchestration/tasks/lifecycle (single helper)
- core/git_authorship → core/projects (only consumer)

Project/Task aggregate OO (the W3.C leftover):
- Task.image_is_old() new method, returns bool | None
- Project.state() renamed from get_state(), accepts gate_commit_provider
- Project.storage_detail() new method
- The underlying free functions stay in domain/project_state.py and
  domain/storage.py as test-friendly internals; only the methods are
  on the public API.

API surface trim (9 entries off lib/api/*):
- EXIT_MANUAL_STEP_NEEDED (defined locally in tui/app.py as a 1-line
  exit-code mirror — the sandbox setup CLI's partial-success signal)
- get_project_state, is_task_image_old (replaced by Project.state(),
  Task.image_is_old())
- ProjectBadge, TaskState (only internal consumers)
- MODE_DISPLAY, ISOLATION_DISPLAY (only internal consumers)
- make_ssh_manager, register_ssh_key (only internal consumers and one
  test reaching past the api surface — test updated to import directly)

Depends on:
- terok-ai/terok-sandbox#354 (ed3fdc6) — adds SelinuxCheckResult back
  to terok_sandbox.__all__ on top of the W4.A trim
- terok-ai/terok-executor#366 (28bae7d) — adds ImageSet back to
  terok_executor.__all__ on top of the W4.B trim and AgentProvider OO
