+++ src/terok/lib/core/projects.py
@@ -35,7 +35,6 @@ from .config import (
     user_presets_dir,
     user_projects_dir,
 )
-from .git_authorship import normalize_git_authorship
 from .project_model import (  # noqa: F401 — re-exported public API
     PresetInfo,
     ProjectConfig,
@@ -48,6 +47,44 @@ logger = logging.getLogger(__name__)
 
 _PROJECT_YML = "project.yml"
 _INSTRUCTIONS_MD = "instructions.md"
+
+# ── Git authorship policy ─────────────────────────────────────────────
+
+DEFAULT_GIT_AUTHORSHIP = "agent-human"
+"""Default Git authorship mode for task containers."""
+
+VALID_GIT_AUTHORSHIP_MODES: tuple[str, ...] = (
+    "agent-human",
+    "human-agent",
+    "agent",
+    "human",
+)
+"""Supported values for ``git.authorship`` in config files."""
+
+
+def normalize_git_authorship(value: object) -> str:
+    """Validate and normalize a ``git.authorship`` config value.
+
+    ``None`` or an empty string fall back to [`DEFAULT_GIT_AUTHORSHIP`][terok.lib.core.projects.DEFAULT_GIT_AUTHORSHIP].
+    Raises [`SystemExit`][SystemExit] for invalid values so project loading can fail
+    with a clear configuration error.
+    """
+    if value is None:
+        return DEFAULT_GIT_AUTHORSHIP
+
+    if not isinstance(value, str):
+        valid = ", ".join(VALID_GIT_AUTHORSHIP_MODES)
+        raise SystemExit(f"Invalid git.authorship value: expected a string.\nValid values: {valid}")
+
+    normalized = value.strip().lower()
+    if not normalized:
+        return DEFAULT_GIT_AUTHORSHIP
+
+    if normalized in VALID_GIT_AUTHORSHIP_MODES:
+        return normalized
+
+    valid = ", ".join(VALID_GIT_AUTHORSHIP_MODES)
+    raise SystemExit(f"Invalid git.authorship value {value!r}.\nValid values: {valid}")
 
 
 def _get_global_git_config(key: str) -> str | None:
