+++ src/terok/lib/integrations/executor.py
@@ -35,11 +35,9 @@ from terok_executor import (  # noqa: F401 — re-exported public API
     TaskStorageInfo,
     acp_socket_is_live,
     agent_doctor_checks,
-    apply_provider_config,
     assemble_container_env,
     authenticate,
     build_base_images,
-    build_headless_command,
     build_project_image,
     build_sidecar_image,
     bundled_default_instructions,
@@ -102,11 +100,9 @@ __all__ = [
     "VAULT_COMMANDS",
     "acp_socket_is_live",
     "agent_doctor_checks",
-    "apply_provider_config",
     "assemble_container_env",
     "authenticate",
     "build_base_images",
-    "build_headless_command",
     "build_project_image",
     "build_sidecar_image",
     "bundled_default_instructions",
