+++ src/terok/lib/integrations/sandbox.py
@@ -23,21 +23,14 @@ imported directly by every layer that needs it.
 from terok_sandbox import (  # noqa: F401 — re-exported public API
     DEFAULT_GUEST_SSHD_PORT,
     DEFAULT_SSH_HOST,
-    DEFAULT_SSH_USER,
-    EXIT_MANUAL_STEP_NEEDED,
-    GATE_COMMANDS,
     SERVICES_TCP_OPTOUT_YAML,
-    SSH_COMMANDS,
     BestEffortLogger,
     CheckVerdict,
-    CommandDef,
-    CommandTree,
     ContainerRuntime,
     CredentialDB,
     DoctorCheck,
     EnvironmentCheck,
     ExecResult,
-    FakeKrunTransport,
     GateAuthNotConfigured,
     GateServerManager,
     GateServerStatus,
@@ -45,12 +38,10 @@ from terok_sandbox import (  # noqa: F401 — re-exported public API
     GitGate,
     Image,
     KrunRuntime,
-    KrunTransport,
     LifecycleHooks,
     NoPassphraseError,
     NullRuntime,
     PodmanRuntime,
-    RawHooksSection,
     RawRunSection,
     RawSSHSection,
     RecoveryStatus,
@@ -62,7 +53,6 @@ from terok_sandbox import (  # noqa: F401 — re-exported public API
     Sharing,
     SSHInitResult,
     SSHManager,
-    TcpEndpoint,
     TcpSSHTransport,
     TokenStore,
     VaultManager,
@@ -81,7 +71,6 @@ from terok_sandbox import (  # noqa: F401 — re-exported public API
     handle_vault_to_keyring,
     installed_versions,
     is_ssh_url,
-    is_systemd_creds_available,
     make_shield,
     needs_setup,
     podman_port_resolver,
@@ -97,8 +86,6 @@ from terok_sandbox import (  # noqa: F401 — re-exported public API
     selinux_install_command,
     selinux_install_script,
     setup_hooks_direct,
-    shield_interactive_session,
-    shield_watch_session,
     stage_line,
     stamp_path,
     state,
@@ -112,19 +99,13 @@ from terok_sandbox import (  # noqa: F401 — re-exported public API
 __all__ = [
     "BestEffortLogger",
     "CheckVerdict",
-    "CommandDef",
-    "CommandTree",
     "ContainerRuntime",
     "CredentialDB",
     "DEFAULT_GUEST_SSHD_PORT",
     "DEFAULT_SSH_HOST",
-    "DEFAULT_SSH_USER",
     "DoctorCheck",
-    "EXIT_MANUAL_STEP_NEEDED",
     "EnvironmentCheck",
     "ExecResult",
-    "FakeKrunTransport",
-    "GATE_COMMANDS",
     "GateAuthNotConfigured",
     "GateServerManager",
     "GateServerStatus",
@@ -132,16 +113,13 @@ __all__ = [
     "GitGate",
     "Image",
     "KrunRuntime",
-    "KrunTransport",
     "LifecycleHooks",
     "NoPassphraseError",
     "NullRuntime",
     "PodmanRuntime",
-    "RawHooksSection",
     "RawRunSection",
     "RawSSHSection",
     "SERVICES_TCP_OPTOUT_YAML",
-    "SSH_COMMANDS",
     "SSHInitResult",
     "SSHManager",
     "Sandbox",
@@ -150,7 +128,6 @@ __all__ = [
     "ServicesMode",
     "SetupVerdict",
     "Sharing",
-    "TcpEndpoint",
     "TcpSSHTransport",
     "TokenStore",
     "VaultManager",
@@ -170,7 +147,6 @@ __all__ = [
     "handle_vault_to_keyring",
     "installed_versions",
     "is_ssh_url",
-    "is_systemd_creds_available",
     "make_shield",
     "needs_setup",
     "podman_port_resolver",
@@ -186,8 +162,6 @@ __all__ = [
     "selinux_install_command",
     "selinux_install_script",
     "setup_hooks_direct",
-    "shield_interactive_session",
-    "shield_watch_session",
     "stage_line",
     "stamp_path",
     "state",
