+++ tests/unit/lib/test_shield.py
@@ -15,18 +15,16 @@ from terok_sandbox import (
     check_environment,
     down,
     make_shield,
-    pre_start,
     run_setup,
     setup_hooks_direct,
     state,
     status,
     up,
 )
-from terok_sandbox.integrations.shield import _BYPASS_WARNING
+from terok_sandbox.integrations.shield import _BYPASS_WARNING, pre_start
 from terok_shield import (
     USER_HOOKS_DIR,
     EnvironmentCheck,
-    NftNotFoundError,
     Shield,
     ShieldMode,
     ShieldNeedsSetup,
@@ -112,20 +110,6 @@ def test_make_shield_maps_config_to_shield_config(
     assert config.audit_enabled is audit_enabled
     assert config.state_dir == MOCK_TASK_DIR / "shield"
     assert config.profiles_dir == MOCK_CONFIG_ROOT / "shield" / "profiles"
-
-
-def test_nft_not_found_is_reexported() -> None:
-    """``NftNotFoundError`` is re-exported from the adapter module."""
-    from terok_sandbox import NftNotFoundError as error_type
-
-    assert error_type is NftNotFoundError
-
-
-def test_shield_state_is_reexported() -> None:
-    """``ShieldState`` is re-exported from the adapter module."""
-    from terok_sandbox import ShieldState as shield_state_type
-
-    assert shield_state_type is ShieldState
 
 
 @pytest.mark.parametrize(
